Add parser stage w/ tests + minor code tree rework
A hand-written parser is necessary in order to work in constant memory with
acceptable performances. The parser feeds the consumer with a stream of
configuration messages.
The speed of the lexing/parsing stage is crucial at startup, or every time
the configuration is changed at runtime, therefore a config file grammar
is picked to allow quick traversal.